move TetheredWorld to its own file
