+++ src/flygym/compose/world/__init__.py
@@ -1,11 +1,12 @@
 from .base_world import BaseWorld
 from .flat_ground import FlatGroundWorld
+from .tethered_world import TetheredWorld
 from .complex_terrain import (
     GappedTerrainWorld,
     BlocksTerrainWorld,
     MixedTerrainWorld,
-    TetheredWorld,
 )
+
 
 __all__ = [
     "BaseWorld",
